# MiqProductFeature - fewer queries to load cache