--- a/app/models/miq_product_feature.rb
+++ b/app/models/miq_product_feature.rb
@@ -5,6 +5,8 @@ class MiqProductFeature < ApplicationRecord
   has_many :miq_product_features_shares
   has_many :shares, :through => :miq_product_features_shares
 
+  virtual_delegate :identifier, :to => :parent, :prefix => true, :allow_nil => true
+
   validates_presence_of   :identifier
   validates_uniqueness_of :identifier
 
@@ -69,12 +71,21 @@ class MiqProductFeature < ApplicationRecord
 
   def self.features
     @feature_cache ||= begin
-      includes(:parent, :children).each_with_object({}) do |f, h|
-        child_idents = f.children.collect(&:identifier)
-        parent_ident = f.parent.identifier if f.parent
+      # create hash with parent identifier and details
+      features = select(:id, :identifier).select(*DETAIL_ATTRS)
+                                         .select(arel_attribute(:parent_identifier).as("parent_identifier"))
+                                         .each_with_object({}) do |f, h|
+        parent_ident = f.parent_identifier
         details      = DETAIL_ATTRS.each_with_object({}) { |a, dh| dh[a] = f.send(a) }
-        h[f.identifier] = {:parent => parent_ident, :children => child_idents, :details => details}
+        h[f.identifier] = {:parent => parent_ident, :children => [], :details => details}
       end
+      # populate the children based upon parent identifier
+      features.each do |identifier, n|
+        if (parent = n[:parent])
+          features[parent][:children] << identifier
+        end
+      end
+      features
     end
   end