--- a/spec/models/miq_product_feature_spec.rb
+++ b/spec/models/miq_product_feature_spec.rb
@@ -168,4 +168,32 @@ describe MiqProductFeature do
       expect(MiqProductFeature.feature_hidden("widget_refresh")).to be_truthy
     end
   end
+
+  describe ".features" do
+    before { MiqProductFeature.instance_variable_set(:@feature_cache, nil) }
+    after  { MiqProductFeature.instance_variable_set(:@feature_cache, nil) }
+
+    #      1
+    #    2    3
+    #        4 5
+    it "populates parent and children" do
+      f1 = FactoryGirl.create(:miq_product_feature, :identifier => "f1", :name => "F1n")
+      FactoryGirl.create(:miq_product_feature, :identifier => "f2", :name => "F2n", :parent_id => f1.id)
+      f3 = FactoryGirl.create(:miq_product_feature, :identifier => "f3", :name => "F3n", :parent_id => f1.id)
+      FactoryGirl.create(:miq_product_feature, :identifier => "f4", :name => "F4n", :parent_id => f3.id)
+      FactoryGirl.create(:miq_product_feature, :identifier => "f5", :name => "F5n", :parent_id => f3.id)
+
+      expect { MiqProductFeature.features }.to match_query_limit_of(1)
+      expect { MiqProductFeature.features }.to match_query_limit_of(0)
+
+      expect(MiqProductFeature.feature_root).to eq("f1")
+      expect(MiqProductFeature.feature_children("f1")).to eq(%w(f2 f3))
+      expect(MiqProductFeature.feature_children("f2")).to eq([])
+      expect(MiqProductFeature.feature_children("f3")).to eq(%w(f4 f5))
+
+      expect(MiqProductFeature.feature_parent("f1")).to be_nil
+      expect(MiqProductFeature.feature_parent("f2")).to eq("f1")
+      expect(MiqProductFeature.feature_parent("f4")).to eq("f3")
+    end
+  end
 end