We load (and cache) product feature identifiers on the first request. This only ever happens once in production mode. This happens often in development mode.

Since this query takes up a page in the logs, it makes collection of metrics difficult.

While this does improve production performance, most of the gains will be seen in development mode.

Summary
----------

|         ms |     bytes |   objects |queries | query (ms) |     rows |log lines|`comments`
|        ---:|       ---:|       ---:|  ---:|       ---:|      ---:| ---|---
|      608.6 | 56,575,039* |   712,271 |    3 |      54.1 |    2,622 |45|`master`
|       85.5 | 5,099,101* |    54,285 |    1 |      19.6 |    1,129 |4|`PR`
| 86% | 91% | 92% | 67% | 64% | 57% | 91% | diff

\* Memory usage does not reflect 143 freed objects. (in both cases)
before
------

|         ms |     bytes |   objects |queries | query (ms) |     rows |`comments`
|        ---:|       ---:|       ---:|  ---:|       ---:|      ---:| ---
|      608.6 | 56,575,039* |   712,271 |    3 |      54.1 |    2,622 |`master`
|       24.1 |           |           |    1 |       2.6 |    1,129 |`SELECT "miq_product_features"  -- [59]`
|        6.2 |           |           |    1 |       2.6 |      365 |`SELECT "miq_product_features"  -- [1908]`
|       23.8 |           |           |    1 |      10.7 |    1,128 |`SELECT "miq_product_features"  -- [5771]`

\* Memory usage does not reflect 143 freed objects. 

If you note, the queries are 7,738 characters (~1 page)

after
----

|         ms |     bytes |   objects |queries | query (ms) |     rows |`comments`
|        ---:|       ---:|       ---:|  ---:|       ---:|      ---:| ---
|       85.5 | 5,099,101* |    54,285 |    1 |      19.6 |    1,129 |`feature_load_query`
|       19.6 |           |           |    1 |       2.9 |    1,129 |`SELECT "miq_product_features" -- [474]`

\* Memory usage does not reflect 143 freed objects. 

If you note, the query takes 474 characters (4 lines)

/cc @chrisarcand thanks for help