@kbrock 2016-09-29
@gtanzillo added tests for the method changed