@chessbyte 2016-10-04
Euwe Backport details:
```sh
$ git log
commit 7ef8e338e37f4f1ee8a6384febe2655824c61d66
Author: Chris Arcand <chrisarcand@users.noreply.github.com>
Date:   Mon Oct 3 15:05:19 2016 -0500

    Merge pull request #11568 from kbrock/feature_load_query
    
    MiqProductFeature - fewer queries to load cache
    (cherry picked from commit 3f13bf87df21f017a042b66a72ebbd5ab4705880)
```